Adding a collision to the collision matrix
On my laptop the tasks could not be executed because of the collision between the surface (a collision object in rviz) and the base_link_inertia of the robot (this is the base of the robot, where it will be screwed on the surface). So here the the robot description will be extracted and a new disable collision will be added to the collision matrix.